spider/pyproject.toml
[project]
name = "spider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "bs4>=0.0.2",
    "lxml>=6.0.2",
    "playwright>=1.58.0",
    "psycopg[binary]>=3.3.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
